Fix up some implicit type conversions into explicit type conversions, and don't build Hammerspoon Tests when running Hammerspoon's Hammerspoon scheme

+++ extensions/drawing/internal.m
@@ -1002,7 +1002,7 @@ static int drawing_setFillGradient(lua_State *L) {
     drawing_t *drawingObject = get_item_arg(L, 1);
     NSColor *startColor = getColorFromStack(L, 2);
     NSColor *endColor = getColorFromStack(L, 3);
-    int angle = lua_tointeger(L, 4);
+    int angle = (int)lua_tointeger(L, 4);
 
     HSDrawingWindow *drawingWindow = (__bridge HSDrawingWindow *)drawingObject->window;
     HSDrawingView *drawingView = (HSDrawingView *)drawingWindow.contentView;
